wasi: backfills constants and TODOs for remaining functions
