+++ wasi/testdata/args.wat
@@ -7,8 +7,10 @@
 ;; this file is put here for now, because this is a temporary file until the parser supports
 ;; the enough syntax, and this file will be embedded in unit test codes after that.
 (module
-  (import "wasi_snapshot_preview1" "args_sizes_get" (func $wasi_args_sizes_get (param i32 i32) (result i32)))
-  (import "wasi_snapshot_preview1" "args_get"       (func $wasi_args_get (param i32 i32) (result i32)))
+  (import "wasi_snapshot_preview1" "args_sizes_get"
+    (func $wasi.args_sizes_get (param $result.argc i32) (param $result.argv_buf_size i32) (result (;errno;) i32)))
+  (import "wasi_snapshot_preview1" "args_get"
+    (func $wasi.args_get (param $argv i32) (param $argv_buf i32) (result (;errno;) i32)))
   (memory 1)  ;; just an arbitrary size big enough for tests
   (export "memory" (memory 0))
   ;; Define wrapper functions instead of just exporting the imported WASI APIS for now
@@ -18,12 +20,12 @@
   (func $args_sizes_get (param i32 i32) (result i32)
         local.get 0
         local.get 1
-        call $wasi_args_sizes_get
+        call $wasi.args_sizes_get
         )
   (func $args_get (param i32 i32) (result i32)
         local.get 0
         local.get 1
-        call $wasi_args_get
+        call $wasi.args_get
         )
   (export "args_sizes_get" (func $args_sizes_get))
   (export "args_get" (func $args_get))
