+++ wasi/wasi.go
@@ -3,7 +3,6 @@ package wasi
 import (
 	"encoding/binary"
 	"errors"
-	"fmt"
 	"io"
 	"io/fs"
 	"math"
@@ -11,12 +10,16 @@ import (
 	"os"
 	"reflect"
 	"time"
-	"unicode/utf8"
 
 	"github.com/tetratelabs/wazero/wasm"
 )
 
-// API is a documentation interface for WASI exported functions in version "wasi_snapshot_preview1"
+// API includes all host functions to export for WASI version "wasi_snapshot_preview1"
+//
+// Note: When translating WASI functions, each result besides Errno is always an uint32 parameter. WebAssembly 1.0 (MVP)
+// can have up to one result, which is already used by Errno. This forces other results to be parameters. A result
+// parameter is a memory offset to write the result to. As memory offsets are uint32, each parameter representing a
+// result is uint32.
 //
 // Note: In WebAssembly 1.0 (MVP), there may be up to one Memory per store, which means the wasm.HostFunctionCallContext
 // Memory is always the wasm.Store Memories index zero: `store.Memories[0].Buffer`
@@ -24,51 +27,18 @@ import (
 // See https://github.com/WebAssembly/WASI/blob/snapshot-01/phases/snapshot/docs.md
 // See https://www.w3.org/TR/wasm-core-1/#memory-instances%E2%91%A0.
 type API interface {
-	// ArgsSizesGet is a WASI function that reads command-line argument data (Args) sizes.
-	//
-	// There are two parameters and an Errno result. Both parameters are offsets in the wasm.HostFunctionCallContext
-	// Memory Buffer to write the corresponding sizes in uint32 little-endian encoding.
-	//
-	// * argc - is the offset to write the argument count to the wasm.MemoryInstance Buffer
-	// * argvBufSize - is the offset to write the null terminated argument length to the wasm.MemoryInstance Buffer
-	//
-	// In WebAssembly 1.0 (MVP) Text format, this signature is:
-	//	(import "wasi_snapshot_preview1" "args_sizes_get"
-	//		(func ($wasi_args_sizes_get (param $argc i32) (param $argv_buf_size i32) (result i32))
-	//
-	// For example, if Args are []string{"a","bc"} and
-	//   ArgsSizesGet parameters argc=1 and argvBufSize=6, we expect `ctx.Memory.Buffer` to contain:
-	//
-	//                    uint32         uint32
-	//                  +--------+     +--------+
-	//                  |        |     |        |
-	//        []byte{?, 2, 0, 0, 0, ?, 5, 0, 0, 0, ?}
-	//           argc --^              ^
-	//         2 args --+              |
-	//                   argvBufSize --|
-	//   len([]byte{'a',0,'b',c',0}) --+
-	//
-	// See ArgsGet
-	// See https://github.com/WebAssembly/WASI/blob/snapshot-01/phases/snapshot/docs.md#args_sizes_get
-	// See https://en.wikipedia.org/wiki/Null-terminated_string
-	ArgsSizesGet(ctx *wasm.HostFunctionCallContext, argc, argvBufSize uint32) Errno
-
-	// ArgsGet is a WASI function that reads command-line argument data (Args).
+	// ArgsGet (FunctionArgsGet) is the WASI function that reads command-line argument data (Args).
 	//
 	// There are two parameters and an Errno result. Both parameters are offsets in the wasm.HostFunctionCallContext
 	// Memory Buffer to write offsets. These are encoded uint32 little-endian.
 	//
 	// * argv - is the offset to begin writing argument offsets to the wasm.MemoryInstance
-	//   * ArgsSizesGet argc * 4 bytes are written to this offset
+	//   * ArgsSizesGet result argc * 4 bytes are written to this offset
 	// * argvBuf - is the offset to write the null terminated arguments to the wasm.MemoryInstance
-	//   * ArgsSizesGet argv_buf_size bytes are written to this offset
-	//
-	// In WebAssembly 1.0 (MVP) Text format, this signature is:
-	//	(import "wasi_snapshot_preview1" "args_get"
-	//		(func ($wasi_args_get (param $argv i32) (param $argv_buf i32) (result i32))
+	//   * ArgsSizesGet result argv_buf_size bytes are written to this offset
 	//
 	// For example, if ArgsSizesGet wrote argc=2 and argvBufSize=5 for arguments: "a" and "bc"
-	//   and ArgsGet parameters argv=7 and argvBuf=1, we expect `ctx.Memory.Buffer` to contain:
+	//   and ArgsGet results argv=7 and argvBuf=1, we expect `ctx.Memory.Buffer` to contain:
 	//
 	//               argvBufSize                argc * 4
 	//            +----------------+     +--------------------+
@@ -83,6 +53,79 @@ type API interface {
 	// See https://github.com/WebAssembly/WASI/blob/snapshot-01/phases/snapshot/docs.md#args_get
 	// See https://en.wikipedia.org/wiki/Null-terminated_string
 	ArgsGet(ctx *wasm.HostFunctionCallContext, argv, argvBuf uint32) Errno
+
+	// ArgsSizesGet (FunctionArgsSizesGet) is a WASI function that reads command-line argument data (Args) sizes.
+	//
+	// There are two parameters and an Errno result. Both parameters are offsets in the wasm.HostFunctionCallContext
+	// Memory Buffer to write the corresponding sizes in uint32 little-endian encoding.
+	//
+	// * argc - is the offset to write the argument count to the wasm.MemoryInstance Buffer
+	// * argvBufSize - is the offset to write the null terminated argument length to the wasm.MemoryInstance Buffer
+	//
+	// For example, if Args are []string{"a","bc"} and
+	//   ArgsSizesGet parameters resultArgc=1 and resultArgvBufSize=6, we expect `ctx.Memory.Buffer` to contain:
+	//
+	//                    uint32         uint32
+	//                  +--------+     +--------+
+	//                  |        |     |        |
+	//        []byte{?, 2, 0, 0, 0, ?, 5, 0, 0, 0, ?}
+	//     resultArgc --^              ^
+	//         2 args --+              |
+	//             resultArgvBufSize --|
+	//   len([]byte{'a',0,'b',c',0}) --+
+	//
+	// See ArgsGet
+	// See https://github.com/WebAssembly/WASI/blob/snapshot-01/phases/snapshot/docs.md#args_sizes_get
+	// See https://en.wikipedia.org/wiki/Null-terminated_string
+	ArgsSizesGet(ctx *wasm.HostFunctionCallContext, resultArgc, resultArgvBufSize uint32) Errno
+
+	// TODO: EnvironGet(ctx *wasm.HostFunctionCallContext, environ, environBuf uint32) Errno
+
+	// TODO: EnvironSizesGet(ctx *wasm.HostFunctionCallContext, resulEnvironc, resultEnvironBufSize uint32) Errno
+
+	// TODO: ClockResGet(ctx *wasm.HostFunctionCallContext, id, resultResolution uint32) Errno
+
+	// TODO: ClockTimeGet(ctx *wasm.HostFunctionCallContext, id uint32, precision uint64, resultTimestamp uint32) Errno
+
+	// TODO: FDAdvise
+	// TODO: FDAllocate
+	// TODO: FDClose
+	// TODO: FDDataSync
+	// TODO: FDFDStatGet
+	// TODO: FDFDStatSetFlags
+	// TODO: FDFDStatSetRights
+	// TODO: FDFilestatGet
+	// TODO: FDFilestatSetSize
+	// TODO: FDFilestatSetTimes
+	// TODO: FDPread
+	// TODO: FDPrestatGet
+	// TODO: FDPrestatDirName
+	// TODO: FDPwrite
+	// TODO: FDRead
+	// TODO: FDReaddir
+	// TODO: FDRenumber
+	// TODO: FDSeek
+	// TODO: FDSync
+	// TODO: FDTell
+	// TODO: FDWrite
+	// TODO: PathCreateDirectory
+	// TODO: PathFilestatGet
+	// TODO: PathFilestatSetTimes
+	// TODO: PathLink
+	// TODO: PathOpen
+	// TODO: PathReadlink
+	// TODO: PathRemoveDirectory
+	// TODO: PathRename
+	// TODO: PathSymlink
+	// TODO: PathUnlinkFile
+	// TODO: PollOneoff
+	// TODO: ProcExit
+	// TODO: ProcRaise
+	// TODO: SchedYield
+	// TODO: RandomGet
+	// TODO: SockRecv
+	// TODO: SockSend
+	// TODO: SockShutdown
 }
 
 const (
@@ -101,65 +144,99 @@ type api struct {
 
 func (w *api) register(store *wasm.Store) (err error) {
 	for _, wasiName := range []string{
-		wasiUnstableName,
+		wasiUnstableName, // TODO: check if there are any signature incompatibility between stable and preview 1
 		wasiSnapshotPreview1Name,
 	} {
-		err = store.AddHostFunction(wasiName, "proc_exit", reflect.ValueOf(proc_exit))
+
+		// Note: these are ordered per spec for consistency
+		// See https://github.com/WebAssembly/WASI/blob/snapshot-01/phases/snapshot/docs.md#functions
+		err = store.AddHostFunction(wasiName, FunctionArgsGet, reflect.ValueOf(w.ArgsGet))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "fd_write", reflect.ValueOf(w.fd_write))
+		err = store.AddHostFunction(wasiName, FunctionArgsSizesGet, reflect.ValueOf(w.ArgsSizesGet))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "environ_sizes_get", reflect.ValueOf(environ_sizes_get))
+		err = store.AddHostFunction(wasiName, FunctionEnvironGet, reflect.ValueOf(environ_get))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "environ_get", reflect.ValueOf(environ_get))
+		err = store.AddHostFunction(wasiName, FunctionEnvironSizesGet, reflect.ValueOf(environ_sizes_get))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "fd_prestat_get", reflect.ValueOf(w.fd_prestat_get))
+		// TODO: FunctionClockResGet
+		err = store.AddHostFunction(wasiName, FunctionClockTimeGet, reflect.ValueOf(w.clock_time_get))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "fd_prestat_dir_name", reflect.ValueOf(w.fd_prestat_dir_name))
+		// TODO: FunctionFDAdvise
+		// TODO: FunctionFDAllocate
+		err = store.AddHostFunction(wasiName, FunctionFDClose, reflect.ValueOf(w.fd_close))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "fd_fdstat_get", reflect.ValueOf(w.fd_fdstat_get))
+		// TODO: FunctionFDDataSync
+		err = store.AddHostFunction(wasiName, FunctionFDFDStatGet, reflect.ValueOf(w.fd_fdstat_get))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "fd_close", reflect.ValueOf(w.fd_close))
+		// TODO: FunctionFDFDStatSetFlags
+		// TODO: FunctionFDFDStatSetRights
+		// TODO: FunctionFDFilestatGet
+		// TODO: FunctionFDFilestatSetSize
+		// TODO: FunctionFDFilestatSetTimes
+		// TODO: FunctionFDPread
+		err = store.AddHostFunction(wasiName, FunctionFDPrestatGet, reflect.ValueOf(w.fd_prestat_get))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "fd_read", reflect.ValueOf(w.fd_read))
+		err = store.AddHostFunction(wasiName, FunctionFDPrestatDirName, reflect.ValueOf(w.fd_prestat_dir_name))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "fd_seek", reflect.ValueOf(w.fd_seek))
+		// TODO: FunctionFDPwrite
+		err = store.AddHostFunction(wasiName, FunctionFDRead, reflect.ValueOf(w.fd_read))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "path_open", reflect.ValueOf(w.path_open))
+		// TODO: FunctionFDReaddir
+		// TODO: FunctionFDRenumber
+		err = store.AddHostFunction(wasiName, FunctionFDSeek, reflect.ValueOf(w.fd_seek))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "args_get", reflect.ValueOf(w.ArgsGet))
+		// TODO: FunctionFDSync
+		// TODO: FunctionFDTell
+		err = store.AddHostFunction(wasiName, FunctionFDWrite, reflect.ValueOf(w.fd_write))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "args_sizes_get", reflect.ValueOf(w.ArgsSizesGet))
+		// TODO: FunctionPathCreateDirectory
+		// TODO: FunctionPathFilestatGet
+		// TODO: FunctionPathFilestatSetTimes
+		// TODO: FunctionPathLink
+		err = store.AddHostFunction(wasiName, FunctionPathOpen, reflect.ValueOf(w.path_open))
 		if err != nil {
 			return err
 		}
-		err = store.AddHostFunction(wasiName, "clock_time_get", reflect.ValueOf(w.clock_time_get))
+		// TODO: FunctionPathReadlink
+		// TODO: FunctionPathRemoveDirectory
+		// TODO: FunctionPathRename
+		// TODO: FunctionPathSymlink
+		// TODO: FunctionPathUnlinkFile
+		// TODO: FunctionPollOneoff
+		err = store.AddHostFunction(wasiName, FunctionProcExit, reflect.ValueOf(proc_exit))
 		if err != nil {
 			return err
 		}
+		// TODO: FunctionProcRaise
+		// TODO: FunctionSchedYield
+		// TODO: FunctionRandomGet
+		// TODO: FunctionSockRecv
+		// TODO: FunctionSockSend
+		// TODO: FunctionSockShutdown
 	}
 	return nil
 }
@@ -188,41 +265,6 @@ func Stderr(writer io.Writer) Option {
 	return func(w *api) {
 		w.stderr = writer
 	}
-}
-
-// nullTerminatedStrings holds null-terminated strings. It ensures that
-// its length and total buffer size don't exceed the max of uint32.
-// nullTerminatedStrings are convenience struct for args_get and environ_get. (environ_get is not implemented yet)
-//
-// A Null-terminated string is a byte string with a NULL suffix ("\x00").
-// See https://en.wikipedia.org/wiki/Null-terminated_string
-type nullTerminatedStrings struct {
-	// nullTerminatedValues are null-terminated values with a NULL suffix.
-	nullTerminatedValues [][]byte
-	totalBufSize         uint32
-}
-
-// newNullTerminatedStrings creates a nullTerminatedStrings from the given string slice. It returns an error
-// if the length or the total buffer size of the result WASIStringArray exceeds the maxBufSize
-func newNullTerminatedStrings(maxBufSize uint32, args ...string) (*nullTerminatedStrings, error) {
-	if len(args) == 0 {
-		return &nullTerminatedStrings{nullTerminatedValues: [][]byte{}}, nil
-	}
-	var strings [][]byte // don't pre-allocate as this function is size bound
-	totalBufSize := uint32(0)
-	for i, arg := range args {
-		if !utf8.ValidString(arg) {
-			return nil, fmt.Errorf("arg[%d] is not a valid UTF-8 string", i)
-		}
-		argLen := uint64(len(arg)) + 1 // + 1 for "\x00"; uint64 in case this one arg is huge
-		nextSize := uint64(totalBufSize) + argLen
-		if nextSize > uint64(maxBufSize) {
-			return nil, fmt.Errorf("arg[%d] will exceed max buffer size %d", i, maxBufSize)
-		}
-		totalBufSize = uint32(nextSize)
-		strings = append(strings, append([]byte(arg), 0))
-	}
-	return &nullTerminatedStrings{nullTerminatedValues: strings, totalBufSize: totalBufSize}, nil
 }
 
 // Args returns an option to give a command-line arguments to the API or errs if the inputs are too large.
@@ -349,8 +391,7 @@ func (w *api) path_open(ctx *wasm.HostFunctionCallContext, fd, dirFlags, pathPtr
 }
 
 func (w *api) fd_seek(ctx *wasm.HostFunctionCallContext, fd uint32, offset uint64, whence uint32, nwrittenPtr uint32) (err Errno) {
-	// not implemented yet
-	return ENOSYS
+	return ENOSYS // TODO: implement
 }
 
 func (w *api) fd_write(ctx *wasm.HostFunctionCallContext, fd uint32, iovsPtr uint32, iovsLen uint32, nwrittenPtr uint32) (err Errno) {
@@ -431,11 +472,11 @@ func (w *api) fd_close(ctx *wasm.HostFunctionCallContext, fd uint32) (err Errno)
 }
 
 // ArgsSizesGet implements API.ArgsSizesGet
-func (w *api) ArgsSizesGet(ctx *wasm.HostFunctionCallContext, argc, argvBufSize uint32) Errno {
-	if !ctx.Memory.PutUint32(argc, uint32(len(w.args.nullTerminatedValues))) {
+func (w *api) ArgsSizesGet(ctx *wasm.HostFunctionCallContext, resultArgc, resultArgvBufSize uint32) Errno {
+	if !ctx.Memory.PutUint32(resultArgc, uint32(len(w.args.nullTerminatedValues))) {
 		return EINVAL
 	}
-	if !ctx.Memory.PutUint32(argvBufSize, w.args.totalBufSize) {
+	if !ctx.Memory.PutUint32(resultArgvBufSize, w.args.totalBufSize) {
 		return EINVAL
 	}
 
@@ -458,16 +499,15 @@ func (w *api) ArgsGet(ctx *wasm.HostFunctionCallContext, argv, argvBuf uint32) E
 }
 
 func proc_exit(*wasm.HostFunctionCallContext, uint32) {
-	// not implemented yet
+	// TODO: implement
 }
 
 func environ_sizes_get(*wasm.HostFunctionCallContext, uint32, uint32) (err Errno) {
-	// not implemented yet
-	return
+	return ENOSYS // TODO: implement
 }
 
 func environ_get(*wasm.HostFunctionCallContext, uint32, uint32) (err Errno) {
-	return
+	return ENOSYS // TODO: implement
 }
 
 // clock_time_get is a WASI API that returns the time value of a clock. Note: This is similar to clock_gettime in POSIX.
